- Renamed classes and file names to be more consistent - Cleaned up pep8 issues
