+++ selenium_tests/lti_emailer/Emailer_Test_Flow.py
@@ -1,7 +1,7 @@
 import unittest
 
 from selenium_tests.lti_emailer.Emailer_Base_Test_Case import EmailerBaseTestCase
-from selenium_tests.lti_emailer.page_objects.emailer_mainpage import MainPage
+from selenium_tests.lti_emailer.page_objects.emailer_mainpage import EmailerMainPage
 
 
 class EmailerTestFlow (EmailerBaseTestCase):
@@ -9,7 +9,7 @@ class EmailerTestFlow (EmailerBaseTestCase):
     def test_is_loaded(self):
         """Check that page is loaded by checking against site title"""
         driver = self.driver
-        page = MainPage(driver)  # instantiate
+        page = EmailerMainPage(driver)  # instantiate
         element = page.get_title()
         page_title = "Course Emailer"
         print "Verifying page title..."
