+++ selenium_tests/lti_emailer/page_objects/emailer_mainpage.py
@@ -1,23 +1,14 @@
 from selenium.webdriver.common.by import By
-from selenium_tests.lti_emailer.page_objects.emailer_base import EmailerBasePage
+from selenium_tests.lti_emailer.page_objects.emailer_base_page import EmailerBasePage
 
-# These are page-specific locators.
-class Locator1(object):
-      _h1_selector = (By.CSS_SELECTOR, "h1")
-      _emailer_list = (By.ID, "email-section-1670")
+# These are page-specific locators on Index/Main page.
+class EmailerMainPageLocators(object):
+
+    _h1_selector = (By.CSS_SELECTOR, "h1")
 
 
-
-class MainPage(EmailerBasePage):
+class EmailerMainPage(EmailerBasePage):
 
     def get_title(self):
-        element = self.find_element(*Locator1._h1_selector)
+        element = self.find_element(*EmailerMainPageLocators._h1_selector)
         return element
-
-    def get_addy(self):
-        email_list = self.find_element(*Locator1._emailer_list)
-        return email_list.text
-
-
-
-
